+++ apps/api/filters/entries.py
@@ -18,8 +18,8 @@ class EntryFilter(django_filters.FilterSet):
     catalog_id = django_filters.UUIDFilter()
     author_id = django_filters.UUIDFilter(method='filter_author_id', label=_("Author"))
     author = django_filters.CharFilter(method='filter_author')
-    category_id = django_filters.UUIDFilter(label=_("Category"))
-    category_term = django_filters.CharFilter(field_name='category__term')
+    category_id = django_filters.UUIDFilter(label=_("Category"), field_name='categories__id')
+    category_term = django_filters.CharFilter(field_name='categories_term')
     language_id = django_filters.UUIDFilter()
     language_code = django_filters.CharFilter(field_name='language__code', label=_("Language"))
     title = django_filters.CharFilter(lookup_expr='unaccent__icontains')
@@ -60,7 +60,7 @@ class EntryFilter(django_filters.FilterSet):
                 title=category.label or category.term,
                 href=f"{self.request.path}?{urlencode(url_params)}",
                 group=_("Category"),
-                count=self.qs.filter(category=category).count(),
+                count=self.qs.filter(categories=category).count(),
                 is_active=self.request.GET.get('category_id') == category.id
             ))