+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.0 : TBD
 
 - **Feature**: Images and thumbnails
+- **Feature**: Entry and categories are m:n
 
 ## 0.3.0 : 2021-08-30